fix: Unable to split again after the first split
After the first split, page selection becomes disabled when attempting
to split again. This issue is because there is no option to re-enable
the page input field.

Furthermore, when splitting a PDF after the first split, the page
selection becomes duplicated, requiring the removal of options in the
page selection before rendering the new page selection.