fix: Duplicate files during merge
When merging PDFs, the list of files to be combined is not reset when
selecting files again. This causes previously selected files (but not
intended for merging) to be included in the new file list.

To address this issue, it is necessary to reset the list of files to
be merged before making a new file selection.